+++ packages/mobile/CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file!
 
 <!-- template-start -->
 
+## 6.5.8 ((4/8/2024, 01:59 PM PST))
+
+This is an artificial version bump with no new change.
+
 ## 6.5.7 (4/8/2024 PST)
 
 #### 🐞 Fixes